+++ orchestrator/src/main.rs
@@ -14,6 +14,7 @@ use planetary_server::DEFAULT_ADDRESS;
 use planetary_server::DEFAULT_PORT;
 use tracing_log::AsTrace as _;
 use tracing_subscriber::EnvFilter;
+use url::Url;
 
 /// The Planetary task orchestrator service.
 #[derive(Parser)]
@@ -29,6 +30,10 @@ pub struct Args {
     /// The pod name of the orchestrator.
     #[clap(long, env)]
     pod_name: String,
+
+    /// The service URL for the orchestrator service.
+    #[clap(long, env)]
+    service_url: Url,
 
     /// The verbosity level.
     #[command(flatten)]
@@ -136,6 +141,7 @@ pub async fn main() -> anyhow::Result<()> {
         .maybe_storage_class(args.storage_class)
         .maybe_transporter_image(args.transporter_image)
         .pod_name(args.pod_name)
+        .service_url(args.service_url)
         .maybe_tasks_namespace(args.tasks_namespace)
         .maybe_transporter_cpu(args.transporter_cpu)
         .maybe_transporter_memory(args.transporter_memory)